Removed warnings from #58

+++ utils/zmq_signaling_server.cpp
@@ -32,7 +32,7 @@ static const MsgType CONTEXT_REQUEST = 6;
 static char * s_recv (zmq::socket_t& socket) {
     zmq::message_t message(256);
     socket.recv(&message);
-    if (message.size() == -1)
+    if (message.size() == static_cast<size_t>(-1))
 	return NULL;
     if (message.size() > 255)
 	static_cast<char*>(message.data())[255] = 0;
